requirements.txt
pandas
requests
schedule
python-dotenv
numpy
